+++ app/project/bootstrappers/databases/Redis.php
@@ -4,13 +4,23 @@
  */
 namespace Project\Bootstrappers\Databases;
 use RDev\Applications\Bootstrappers\Bootstrapper;
+use RDev\Applications\Bootstrappers\ILazyBootstrapper;
 use RDev\IoC\IContainer;
+use RDev\Redis\IRedis;
 use RDev\Redis\RDevPHPRedis;
 use RDev\Redis\Server;
 use RDev\Redis\TypeMapper;
 
-class Redis extends Bootstrapper
+class Redis extends Bootstrapper implements ILazyBootstrapper
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function getBoundClasses()
+    {
+        return [IRedis::class];
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -24,6 +34,6 @@ class Redis extends Bootstrapper
             ),
             new TypeMapper()
         );
-        $container->bind("RDev\\Redis\\IRedis", $redis);
+        $container->bind(IRedis::class, $redis);
     }
 }