+++ bootstrap/console/start.php
@@ -2,6 +2,7 @@
 /**
  * Boots up our application with a console kernel
  */
+use RDev\Applications\Bootstrappers\IO\BootstrapperIO;
 use RDev\Console\Commands\CommandCollection;
 use RDev\Console\Commands\Compilers\ICompiler;
 use RDev\Console\Requests\Parsers\IParser;
@@ -11,10 +12,22 @@ require_once __DIR__ . "/../start.php";
 
 /**
  * ----------------------------------------------------------
+ * Finish setting up the bootstrappers
+ * ----------------------------------------------------------
+ */
+$bootstrapperIO->registerBootstrapperClasses(require $application->getPaths()["configs"] . "/console/bootstrappers.php");
+$application->registerPreStartTask(function() use ($bootstrapperDispatcher, &$bootstrapperIO)
+{
+    $bootstrapperDispatcher->dispatch(
+        $bootstrapperIO->read(BootstrapperIO::CACHED_CONSOLE_BOOTSTRAPPER_REGISTRY_FILE_NAME)
+    );
+});
+
+/**
+ * ----------------------------------------------------------
  * Let's get started
  * ----------------------------------------------------------
  */
-$application->registerBootstrappers(require $application->getPaths()["configs"] . "/console/bootstrappers.php");
 $statusCode = $application->start(function () use ($application)
 {
     global $argv;