+++ bootstrap/http/start.php
@@ -2,6 +2,7 @@
 /**
  * Boots up our application with an HTTP kernel
  */
+use RDev\Applications\Bootstrappers\IO\BootstrapperIO;
 use RDev\Framework\HTTP\Kernel;
 use RDev\HTTP\Requests\Request;
 use RDev\Routing\Router;
@@ -10,10 +11,22 @@ require_once __DIR__ . "/../start.php";
 
 /**
  * ----------------------------------------------------------
+ * Finish setting up the bootstrappers
+ * ----------------------------------------------------------
+ */
+$bootstrapperIO->registerBootstrapperClasses(require $application->getPaths()["configs"] . "/http/bootstrappers.php");
+$application->registerPreStartTask(function() use ($bootstrapperDispatcher, &$bootstrapperIO)
+{
+    $bootstrapperDispatcher->dispatch(
+        $bootstrapperIO->read(BootstrapperIO::CACHED_HTTP_BOOTSTRAPPER_REGISTRY_FILE_NAME)
+    );
+});
+
+/**
+ * ----------------------------------------------------------
  * Let's get started
  * ----------------------------------------------------------
  */
-$application->registerBootstrappers(require $application->getPaths()["configs"] . "/http/bootstrappers.php");
 $application->start(function() use ($application)
 {
     /**