+++ configs/bootstrappers.php
@@ -2,7 +2,6 @@
 /**
  * Defines the list of bootstrapper classes to load
  */
-use Project\Bootstrappers\Authentication\Credentials;
 use Project\Bootstrappers\Databases\SQL;
 use Project\Bootstrappers\ORM\UnitOfWork;
 use RDev\Framework\Bootstrappers\Cryptography\Cryptography;
@@ -29,7 +28,6 @@ return [
      * List any console bootstrappers you'd like here
      * To enable Redis, add "Project\\Bootstrappers\\Databases\\Redis"
      */
-    Credentials::class,
     SQL::class,
     UnitOfWork::class
 ];