+++ tests/app/project/console/ApplicationTestCase.php
@@ -4,6 +4,8 @@
  */
 namespace Project\Console;
 use RDev\Applications\Application;
+use RDev\Applications\Bootstrappers\Dispatchers\IDispatcher;
+use RDev\Applications\Bootstrappers\IO\BootstrapperIO;
 use RDev\Framework\Tests\Console\ApplicationTestCase as BaseTestCase;
 
 class ApplicationTestCase extends BaseTestCase
@@ -14,8 +16,17 @@ class ApplicationTestCase extends BaseTestCase
     protected function setApplication()
     {
         /** @var Application $application */
+        /** @var BootstrapperIO $bootstrapperIO */
+        /** @var IDispatcher $bootstrapperDispatcher */
         require __DIR__ . "/../../../../bootstrap/start.php";
-        $application->registerBootstrappers(require $application->getPaths()["configs"] . "/console/bootstrappers.php");
+        $consoleBootstrapperClasses = require $application->getPaths()["configs"] . "/console/bootstrappers.php";
+        $bootstrapperIO->registerBootstrapperClasses($consoleBootstrapperClasses);
+        $application->registerPreStartTask(function() use ($bootstrapperDispatcher, &$bootstrapperIO)
+        {
+            $bootstrapperDispatcher->dispatch(
+                $bootstrapperIO->read(BootstrapperIO::CACHED_CONSOLE_BOOTSTRAPPER_REGISTRY_FILE_NAME)
+            );
+        });
         $this->application = $application;
     }
 }